requirements.txt
requests
beautifulsoup4
pandas
reportlab
python-dateutil
flask
flask-cors
gunicorn
python-dotenv
